+++ crates/orchestrator/src/jobs/types.rs
@@ -89,26 +89,19 @@ pub enum JobType {
 pub enum JobStatus {
     /// An acknowledgement that the job has been received by the
     /// orchestrator and is waiting to be processed
-    #[strum(to_string = "Created")]
     Created,
     /// Some system has taken a lock over the job for processing and no
     /// other system to process the job
-    #[strum(to_string = "Locked for Processing")]
     LockedForProcessing,
     /// The job has been processed and is pending verification
-    #[strum(to_string = "Pending Verification")]
     PendingVerification,
     /// The job has been processed and verified. No other actions needs to be taken
-    #[strum(to_string = "Completed")]
     Completed,
     /// The job was processed but the was unable to be verified under the given time
-    #[strum(to_string = "Verification Timeout")]
     VerificationTimeout,
     /// The job failed processing
-    #[strum(to_string = "Verification Failed")]
     VerificationFailed,
     /// The job failed completing
-    #[strum(to_string = "Failed")]
     Failed,
 }